edits summarized in description
correcting typos, adding back a few of Brian's updates that were missed in the last merge, and updated the pctymle plot to be untransformed (we look at untransformed pctymle in our final specifications so this was updated to maintain consistency throughout the report). Also renamed files to the final format!